+++ src/components/ui/label.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import * as LabelPrimitive from '@radix-ui/react-label';
+import { cva } from 'class-variance-authority';
+import { cn } from '@/lib/utils';
+
+const labelStyles = cva(
+  'text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70'
+);
+
+export const Label = React.forwardRef(({ className, ...props }, ref) => (
+  <LabelPrimitive.Root
+    ref={ref}
+    className={cn(labelStyles(), className)} 
+    {...props}
+  />
+));
+
+Label.displayName = 'Label';
